utils/restutil/view: refactor per resource 'view' package into shared utils/restutil/view
Changelog: None
