+++ utils/restutil/view/view_external_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/mendersoftware/go-lib-micro/log"
 	"github.com/stretchr/testify/assert"
 
-	. "github.com/mendersoftware/deployments/resources/images/view"
+	. "github.com/mendersoftware/deployments/utils/restutil/view"
 )
 
 func TestRenderPost(t *testing.T) {
